Tests/LibWeb/Text/expected/css/animated-post-compute-adjustments.txt

overflow-x: visible
overflow-y: visible
display: inline
position: static
